Initial renaming from pLaunchController to pLaunchControl. It does not include updates to documentation. Relates to #6
