+++ .gitignore
@@ -32,8 +32,7 @@
 # .idea/*.iml
 # .idea/modules
 
-# CMake
-cmake-build-*/
+
 
 # Mongo Explorer plugin
 .idea/**/mongoSettings.xml
@@ -43,24 +42,4 @@ cmake-build-*/
 
 # IntelliJ
 out/
-
-# mpeltonen/sbt-idea plugin
-.idea_modules/
-
-# JIRA plugin
-atlassian-ide-plugin.xml
-
-# Cursive Clojure plugin
-.idea/replstate.xml
-
-# Crashlytics plugin (for Android Studio and IntelliJ)
-com_crashlytics_export_strings.xml
-crashlytics.properties
-crashlytics-build.properties
-fabric.properties
-
-# Editor-based Rest Client
-.idea/httpRequests
-
-# Android studio 3.1+ serialized cache file
-.idea/caches/build_file_checksums.ser
+/out/**
