+++ examples/Vertices/Vertices.pde
@@ -1,5 +1,5 @@
 /**
- * Demonstrates pLaunchController by using the pads to switch between three ways of drawing a cylinder-like geometry.
+ * Demonstrates pLaunchControl by using the pads to switch between three ways of drawing a cylinder-like geometry.
  * In this examples, the pads work as "radio-buttons", meaning only one can be activated. Note that this demo uses only pads 1 to 3.
  * 
  * Credits to original code: Processing example "Vertices", by Simon Greenwold.
@@ -10,7 +10,7 @@
  *  number of sides drawn is variable.
  */
  
-import pLaunchController.*;
+import pLaunchControl.*;
 
 LaunchController controller;
